+++ activejob/lib/active_job/railtie.rb
@@ -29,25 +29,6 @@ module ActiveJob
       ActiveSupport.on_load(:active_job) do
         ActiveSupport.on_load(:active_record) do
           ActiveJob::Base.include EnqueueAfterTransactionCommit
-
-          if app.config.active_job.key?(:enqueue_after_transaction_commit)
-            ActiveJob.deprecator.warn(<<~MSG.squish)
-              `config.active_job.enqueue_after_transaction_commit` is deprecated and will be removed in Rails 8.1.
-              This configuration can still be set on individual jobs using `self.enqueue_after_transaction_commit=`,
-              but due the nature of this behavior, it is not recommended to be set globally.
-            MSG
-
-            value = case app.config.active_job.enqueue_after_transaction_commit
-            when :always
-              true
-            when :never
-              false
-            else
-              false
-            end
-
-            ActiveJob::Base.enqueue_after_transaction_commit = value
-          end
         end
       end
     end
@@ -78,6 +59,7 @@ module ActiveJob
         options = options.except(
           :log_query_tags_around_perform,
           :custom_serializers,
+          # This config can't be applied globally, so we need to remove otherwise it will be applied to `ActiveJob::Base`.
           :enqueue_after_transaction_commit
         )